chore: separate client queries
